Initial
First pass purely treated middleware as tokio services (meaning only
"around" middleware existed), and we just provided methods for
constructing middleware stacks. The main concern with this design was
that the path of least resistance involved boxing (see
https://gist.github.com/sgrif/808c69935e840e6ad998ffd29a7963f2). I think
the solution is to provide proper sugar for before and after middleware.

This design has the issue of not allowing access to self in after. I
can't see an easy way around this with the trait form, since we'd need
the lifetime for the future. We could work around this with a copy
constraint but I don't like that either. Going to explore an alternative
design focused on closures instead.