Provide access to `&self` in after middleware
After some lengthy discussions I realized that we can't have a pure
reference here, as we can't rely on stack data in futures at the moment.
This might change in the future but for now any config data will need to
be in an `Arc`.

My main issue with this design is that we're cloning *all* of self. I'd
rather have the Arc be opaque behind the hood, and not require clone on
the middlewares further down the stack.